Mockups and Schema Updates
Moved SiteConditionRule and PlantAttribute to their own file
Added mockup data in routes for forestType, littleForests, and plants
