+++ api/src/database/schema/forestType.ts
@@ -8,17 +8,7 @@ images: string[]
 */
 import { getModelForClass, prop, ReturnModelType } from "@typegoose/typegoose";
 import { Plant } from "./plants";
-
-export class SiteConditionRule{
-    @prop()
-    public name!: string;  //name of rule
-
-    @prop({type: String, required: true, default: []})
-    public allowedValues!: string[];  //values allowed in the rule
-
-    @prop({type: String, required: true, default: []})
-    public images!: string[];
-}
+import { SiteConditionRule } from "./siteConditionRule";
 
 
 export class ForestType{
@@ -41,11 +31,6 @@ export class ForestType{
     active!: true;
 
 }
-
-export const SiteConditionRules:ReturnModelType<
-    typeof SiteConditionRule,
-    {}
-> = getModelForClass(SiteConditionRule, {schemaOptions: {collection: "siteConditionRules"}});
 
 export const ForestTypes: ReturnModelType<
     typeof ForestType,
