+++ api/src/database/schema/littleForests.ts
@@ -11,7 +11,8 @@ images: String[]
 */
 
 import { getModelForClass, prop, ReturnModelType } from "@typegoose/typegoose";
-import { ForestType, SiteConditionRule } from "./forestType";
+import { ForestType} from "./forestType";
+import { SiteConditionRule } from "./siteConditionRule";
 import { Plant } from "./plants";
 
 
@@ -28,8 +29,8 @@ export class LittleForest{
     @prop({type: [Plant, Number], required: true, default: []})
     public plants!: [Plant, number][];  //plant and how many are ordered
 
-    @prop()
-    public isPublic!: false;
+    @prop({type: Boolean, required: true, default: false})
+    public isPublic!: boolean;
 
     @prop({type: String, required: true, default: []})
     public images!: string[];
