+++ api/src/routes/forestTypes.ts
@@ -1,13 +1,18 @@
 import { Router, Request, Response, NextFunction } from "express";
 import { getAllForestTypes, getForestTypeByID } from "src/database/forestTypes";
+import { ForestType } from "src/database/schema/forestType";
 
 const router: Router = Router();
+const f1: ForestType = {name: 'forestType1', siteConditions: [], layersRatio: [[1, 1]], domSpecies: [], images: [], active: true};
+const f2: ForestType = {name: 'forestType2', siteConditions: [], layersRatio: [[2, 2]], domSpecies: [], images: [], active: true};
+const f3: ForestType = {name: 'forestType3', siteConditions: [], layersRatio: [[3, 3]], domSpecies: [], images: [], active: true};
 
 router.get("/forests/", (req: Request, res: Response) => {
-	res.send(getAllForestTypes());
+	// if not null, then:
+	res.send({f1, f2, f3});
 });
 router.get("/forests/:id", (req: Request, res: Response) => {
-	res.send(getForestTypeByID(req.params.id));
+	res.send(f1);
 });
 
 router.get(
